Split Recalculate view into navigation and content components
